Fix port exhaustion in parallel test execution using a custom pool
In .NET Framework TFMs, when running thousands of tests in parallel over HTTP, we can exhaust ports due to too many created connections. So we have to pool a reasonable number of them to unblock execution.The default pool implementation is still suspect, as it creates an HttpClient on every single request, obviating the benefit of a factory.